backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.0
python-dotenv==1.0.0
sqlalchemy==2.0.0
pyodbc==5.0.0
requests==2.31.0
pytest==7.4.0
pytest-cov==4.1.0
